Refactor OLS predict with for loops for clarity
